+++ .eslintrc.js
@@ -125,8 +125,6 @@ module.exports = {
         Game: true,
         UserSpectatingGame: true,
         GameStateRow:true,
-        GameState:true,
-        Player: true
       },
       globals: sailsGlobals,
       rules: {
