+++ tests/e2e/support/helpers.js
@@ -664,7 +664,7 @@ export function cardsArraysMatch(res, fixture) {
  *   p1FaceCards: {suit: number, rank: number}[],
  * }
  */
-export function assertResGameStateixture(res,fixture) {
+export function assertResGameStateFixture(res,fixture) {
 
     // Player 0
     expect(cardListsMatch(res.players[0].hand, fixture.p0Hand)).to.eq(
@@ -711,3 +711,54 @@ export function assertResGameStateixture(res,fixture) {
       )} did not match ficture: ${printCardList(fixture.p1FaceCards)}`,
     );
 }
+
+/**
+ * @param stringFormat:
+ * {
+ * 	 p0Hand: string[],
+ *   p0Points: string[],
+ *   p0FaceCards: string[],
+ *   p1Hand: string[],
+ *   p1Points: string[],
+ *   p1FaceCards: string[],
+ * }
+ * @param res{
+ * 	 p0Hand: {suit: number, rank: number}[],
+ *   p0Points: {suit: number, rank: number}[],
+ *   p0FaceCards: {suit: number, rank: number}[],
+ *   p1Hand: {suit: number, rank: number}[],
+ *   p1Points: {suit: number, rank: number}[],
+ *   p1FaceCards: {suit: number, rank: number}[],
+ * }
+ */
+export function assertResGameStateString(res, stringFormat) {
+  expect(cardsArraysMatch(res.p0Hand, stringFormat.p0Hand)).to.eq(
+    true,
+    `GameStateRow P0Hand should match ${stringFormat.p0Hand}, 
+    but actual: ${res.p0Hand} did not match fixture`
+  );
+  expect( cardsArraysMatch(res.p0FaceCards, stringFormat.p0FaceCards)).to.eq( 
+    true,
+    `GameStateRow p0FaceCards should match ${stringFormat.p0FaceCards}, 
+    but actual: ${res.p0FaceCards} did not match ficture`);
+
+  expect( cardsArraysMatch(res.p0Points, stringFormat.p0Points)).to.eq( 
+    true,
+    `GameStateRow p0Points should match ${stringFormat.p0Points}, 
+    but actual: ${res.p0Points} did not match ficture`);
+
+  expect( cardsArraysMatch(res.p1Hand, stringFormat.p1Hand) ).to.eq( 
+    true,
+    `GameStateRow p1Hand should match ${stringFormat.p1Hand}, 
+    but actual: ${res.p1Hand} did not match ficture`);
+
+  expect( cardsArraysMatch(res.p1FaceCards, stringFormat.p1FaceCards)).to.eq( 
+    true,
+    `GameStateRow p1FaceCards should match ${stringFormat.p1FaceCards}, 
+    but actual: ${res.p1FaceCards} did not match ficture`);
+
+  expect( cardsArraysMatch(res.p1Points, stringFormat.p1Points)).to.eq( 
+    true,
+    `GameStateRow p1Points should match ${stringFormat.p1Points}, 
+    but actual: ${res.p1Points} did not match ficture`);
+}
